+++ modules/tinymce/src/plugins/preview/main/ts/ui/Dialog.ts
@@ -15,7 +15,8 @@ export const open = (editor: Editor): void => {
         {
           name: 'preview',
           type: 'iframe',
-          sandboxed: true
+          sandboxed: true,
+          opaque: true
         }
       ]
     },