+++ modules/tinymce/src/themes/silver/main/ts/ui/dialog/IFrame.ts
@@ -34,11 +34,13 @@ const getDynamicSource = (initialData: Optional<string>): IFrameSourcing => {
 
 const renderIFrame = (spec: IframeSpec, providersBackstage: UiFactoryBackstageProviders, initialData: Optional<string>) => {
   const isSandbox = spec.sandboxed;
+  const isOpaque = spec.opaque;
 
   const attributes = {
     ...spec.label.map<{ title?: string }>((title) => ({ title })).getOr({}),
     ...initialData.map((html) => ({ srcdoc: html })).getOr({}),
-    ...isSandbox ? { sandbox: 'allow-scripts allow-same-origin' } : { }
+    ...isSandbox ? { sandbox: 'allow-scripts allow-same-origin' } : { },
+    ...isOpaque ? { class: 'tox-dialog__iframe tox-dialog__iframe--opaque' } : { class: 'tox-dialog__iframe' }
   };
 
   const sourcing = getDynamicSource(initialData);